launch de-coupling
