+++ lbr_bringup/launch/move_group.launch.py
@@ -3,7 +3,7 @@ from typing import List
 from launch import LaunchContext, LaunchDescription, LaunchDescriptionEntity
 from launch.actions import OpaqueFunction
 from launch.substitutions import LaunchConfiguration
-from lbr_bringup.description import LBRDescriptionMixin, RVizMixin
+from lbr_bringup.description import LBRDescriptionMixin
 from lbr_bringup.move_group import LBRMoveGroupMixin
 
 
@@ -40,28 +40,6 @@ def launch_setup(context: LaunchContext) -> List[LaunchDescriptionEntity]:
             namespace=robot_name,
         )
     )
-
-    # RViz
-    rviz = RVizMixin.node_rviz(
-        rviz_config_pkg=f"{model}_moveit_config",
-        rviz_config="config/moveit.rviz",
-        parameters=LBRMoveGroupMixin.params_rviz(
-            moveit_configs=moveit_configs_builder.to_moveit_configs()
-        )
-        + [{"use_sim_time": use_sim_time}],
-        remappings=[
-            ("display_planned_path", robot_name + "/display_planned_path"),
-            ("joint_states", robot_name + "/joint_states"),
-            ("monitored_planning_scene", robot_name + "/monitored_planning_scene"),
-            ("planning_scene", robot_name + "/planning_scene"),
-            ("planning_scene_world", robot_name + "/planning_scene_world"),
-            ("robot_description", robot_name + "/robot_description"),
-            ("robot_description_semantic", robot_name + "/robot_description_semantic"),
-        ],
-    )
-
-    ld.add_action(rviz)
-
     return ld.entities
