+++ lbr_bringup/lbr_bringup/gazebo.py
@@ -1,0 +1,49 @@
+from typing import List, Optional, Union
+
+from launch.actions import IncludeLaunchDescription
+from launch.launch_description_sources import PythonLaunchDescriptionSource
+from launch.substitutions import LaunchConfiguration, PathJoinSubstitution
+from launch_ros.actions import Node
+from launch_ros.substitutions import FindPackageShare
+
+
+class GazeboMixin:
+    @staticmethod
+    def include_gazebo(**kwargs) -> IncludeLaunchDescription:
+        return IncludeLaunchDescription(
+            PythonLaunchDescriptionSource(
+                PathJoinSubstitution(
+                    [
+                        FindPackageShare("gazebo_ros"),
+                        "launch",
+                        "gazebo.launch.py",
+                    ]
+                )
+            ),
+            **kwargs,
+        )
+
+    @staticmethod
+    def node_spawn_entity(
+        robot_name: Optional[Union[LaunchConfiguration, str]] = LaunchConfiguration(
+            "robot_name", default="lbr"
+        ),
+        tf: List[float] = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
+        **kwargs,
+    ) -> Node:
+        label = ["-x", "-y", "-z", "-R", "-P", "-Y"]
+        tf = [str(x) for x in tf]
+        return Node(
+            package="gazebo_ros",
+            executable="spawn_entity.py",
+            arguments=[
+                "-topic",
+                "robot_description",
+                "-entity",
+                robot_name,
+            ]
+            + [item for pair in zip(label, tf) for item in pair],
+            output="screen",
+            namespace=robot_name,
+            **kwargs,
+        )
