+++ lbr_demos/lbr_moveit_cpp/launch/hello_moveit.launch.py
@@ -3,15 +3,19 @@ from typing import List
 from launch import LaunchContext, LaunchDescription, LaunchDescriptionEntity
 from launch.actions import OpaqueFunction
 from launch.substitutions import LaunchConfiguration
-from launch_mixins.lbr_bringup import LBRMoveGroupMixin
-from launch_mixins.lbr_description import LBRDescriptionMixin
 from launch_ros.actions import Node
+from lbr_bringup.description import LBRDescriptionMixin
+from lbr_bringup.move_group import LBRMoveGroupMixin
 
 
 def launch_setup(context: LaunchContext) -> List[LaunchDescriptionEntity]:
     ld = LaunchDescription()
 
     model = LaunchConfiguration("model").perform(context)
+    mode = LaunchConfiguration("mode").perform(context)
+    use_sim_time = False
+    if mode == "gazebo":
+        use_sim_time = True
 
     # generate moveit configs
     moveit_configs = LBRMoveGroupMixin.moveit_configs_builder(
@@ -26,7 +30,7 @@ def launch_setup(context: LaunchContext) -> List[LaunchDescriptionEntity]:
             executable="hello_moveit",
             parameters=[
                 moveit_configs.to_dict(),
-                {"use_sim_time": LaunchConfiguration("sim")},
+                {"use_sim_time": use_sim_time},
                 LBRDescriptionMixin.param_robot_name(),
             ],
         )
@@ -39,7 +43,7 @@ def generate_launch_description() -> LaunchDescription:
 
     ld.add_action(LBRDescriptionMixin.arg_model())
     ld.add_action(LBRDescriptionMixin.arg_robot_name())
-    ld.add_action(LBRDescriptionMixin.arg_sim())
+    ld.add_action(LBRDescriptionMixin.arg_mode())
 
     ld.add_action(OpaqueFunction(function=launch_setup))
